[modify] 調整共用 function
