+++ app/Services/LineBot/BaseService.php
@@ -2,6 +2,8 @@
 
 namespace App\Services\LineBot;
 
+use App\Models\GroupAdmin;
+use App\Models\GroupConfig;
 use Illuminate\Support\Facades\Artisan;
 use ReflectionClass;
 
@@ -75,5 +77,41 @@ class BaseService
         }
 
         return $options;
+    }
+
+    /**
+     * 群組設定
+     *
+     * @return GroupConfig
+     */
+    public static function groupConfig($groupId)
+    {
+        $find = GroupConfig::where('group_id', $groupId)->first();
+
+        if ($find == null) {
+            $find              = new GroupConfig();
+            $find->group_id    = $groupId;
+            $find->silent_mode = true;
+            $find->save();
+        }
+
+        return $find;
+    }
+
+    /**
+     * 該群組的管理員
+     *
+     * @param string $groupId // 群組 id
+     * @return GroupAdmin|null
+     */
+    public static function groupAdmin(string $groupId)
+    {
+        return GroupAdmin::where('group_id', function ($sub) use ($groupId) {
+            return $sub->select('id')
+                ->from('group_config')
+                ->where('group_id', $groupId);
+        })
+            ->where('is_sidekick', false)
+            ->first();
     }
 }
